ble_presence_agent/requirements.txt
bleak==0.20.2
requests==2.28.1
python-dotenv==0.21.0
aiofiles
asyncio
requests
scikit-learn
xgboost
joblib
python-dotenv
pandas
numpy
